fixed migration

+++ app/database/migrations/2014_02_04_205909_drop_recipes_categories_table.php
@@ -12,7 +12,9 @@ class DropRecipesCategoriesTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::drop('recipes_categories');
+		if(Schema::hasTable('recipes_categories')){
+			Schema::drop('recipes_categories');
+		}		
 	}
 
 	/**
